aarogya-backend/requirements.txt
Flask==3.1.0
Flask-PyMongo==3.0.1
google-generativeai==0.8.4
groq==0.18.0
# mediapipe==0.10.21
# numpy==1.26.4
# opencv-python==4.11.0.86
PyJWT==2.10.1
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
Werkzeug==3.1.3
pillow
gunicorn